requirements.txt
# Core dependencies
dspy>=2.6.0
openai>=1.0.0
anthropic>=0.18.0
nltk>=3.8
requests>=2.25.0
flask>=2.0.0
pydantic>=2.0.0
python-dotenv>=0.19.0

# Metrics and evaluation
rouge-score>=0.1.2
rouge>=1.0.0
bert-score>=0.3.13
langdetect>=1.0.9

# Data processing
datasets>=2.14.0
pandas>=1.5.0
numpy>=1.21.0

# Utilities
backoff>=2.2.0
pyyaml>=6.0
colorama>=0.4.6 
ujson>=5.0.0
litellm>=2.0.0